requirements.txt
pytest>=7.1.0
matplotlib==3.5.3
seaborn==0.11.2
plotly==5.10.0
pyyaml>=5.3
pandas>=2.1.3
openmeteo_requests==1.2.0
requests_cache==1.2.0
retry_requests==2.0.0
